+++ src/features/create_layer/atoms/editableLayersLegendsAndSources.ts
@@ -5,6 +5,8 @@ import {
   convertDetailsToSource,
   convertDetailsToLegends,
 } from '~core/logical_layers/utils/convert';
+import { forceRun } from '~utils/atoms/forceRun';
+import { createLayerController } from '../control';
 import { editableLayersDetailsResourceAtom } from './editableLayersDetailsResource';
 import type { Action } from '@reatom/core';
 
@@ -72,3 +74,5 @@ export const editableLayersLegendsAndSources = createAtom(
   },
   'editableLayersLegendsAndSources',
 );
+
+createLayerController.onInit(() => forceRun(editableLayersLegendsAndSources));
