+++ src/features/create_layer/components/EditLayerPanel/EditLayerPanel.tsx
@@ -6,9 +6,9 @@ import { i18n } from '~core/localization';
 import { createStateMap } from '~utils/atoms';
 import { LoadingSpinner } from '~components/LoadingSpinner/LoadingSpinner';
 import { ErrorMessage } from '~components/ErrorMessage/ErrorMessage';
-import { editTargetAtom } from '~features/create_layer/atoms/editTarget';
 import { IS_MOBILE_QUERY, useMediaQuery } from '~utils/hooks/useMediaQuery';
-import { createLayerController } from '../../index';
+import { editTargetAtom } from '../../atoms/editTarget';
+import { createLayerController } from '../../control';
 import { EditLayerForm } from '../../components/EditLayerForm/EditLayerForm';
 import { editableLayerControllerAtom } from '../../atoms/editableLayerController';
 import { EditTargets } from '../../constants';
